src/requirements.txt
cloudshell-shell-core~=6.0
cloudshell-cp-core~=2.5.0
cloudshell-cp-gcp~=1.0
